Landing Page에 Footer 추가, 더즐 화면 분리
