+++ client/src/components/views/LandingPage/Sections/LandingPage.css
@@ -1,74 +1,6 @@
 /* Landing Page 전체 컨테이너 */
-#LandingPage {
-  width: 100%;
-  height: 100%;
-
-  margin-left: 6.8%;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-
-  box-sizing: border-box;
-  padding-bottom: 11.8vh;
-}
-
-/* 더 즐겁게, ~ 하라 */
-.MainText {
-  margin: 0;
-
-  display: flex;
-  flex-direction: column;
-
-  font-size: 6rem;
-  font-weight: 900;
-}
-
-/* puzzle gif */
-.PuzzleGif {
-  position: absolute;
-  right: 0;
-}
-
-.PuzzleGif img {
-  width: 50vw;
-}
-
-/* SNS Link */
-.SNSContainer {
-  display: flex;
-
-  position: absolute;
-  bottom: 9.3vh;
-
-  font-size: 1.1rem;
-  font-weight: bold;
-}
-/* text link */
-.SNSContainer a {
-  text-decoration: none;
-  color: #2d2d2d;
-  margin-right: 18px;
-}
-/* text link & arrow image */
-.Instagram,
-.Facebook {
-  display: flex;
-  align-items: center;
-  margin-right: 3.6vw;
-}
-
-/* 더 즐겁게, TOGETHER 하라 */
-.Together {
-  color: #00f5b7;
-}
-
-/* ~와 하라 사이 마진 */
-.Text2 {
-  margin-right: 10px;
-}
-
-/* 더 즐겁게, 와 ~ 하라 간격 줄이기 */
-.Text2Container {
-  margin-top: -20px;
+.LandingPageContainer {
+  height: 90.7vh;
+  overflow-y: scroll;
+  overflow-x: hidden;
 }
