+++ client/src/utils/Common.css
@@ -10,7 +10,7 @@ body,
 }
 
 #Container {
-  height: 90.7%;
+  height: 90.7vh;
   width: 100%;
 
   font-family: "Noto Sans KR", sans-serif;
